Remove deprecated effects and refactor the effect loader
