+++ backend/effects/builtInEffectLoader.js
@@ -3,124 +3,56 @@
 const effectManager = require("./effectManager");
 
 exports.loadEffects = () => {
-    // get effect definitions
-    const playSoundEffect = require("./builtin/playSound");
-    const chatEffect = require("./builtin/chat");
-    const chatFeedAlert = require("./builtin/chat-feed-alert");
-    const api = require("./builtin/api");
-    const celebration = require("./builtin/celebration");
-    const clips = require('./builtin/clips');
-    const dice = require("./builtin/dice");
-    const fileWriter = require("./builtin/fileWriter");
-    const html = require("./builtin/html");
-    const playVideo = require("./builtin/playVideo");
-    const showEvents = require("./builtin/showEvents");
-    const controlEmulation = require("./builtin/controlEmulation");
-    const showImage = require("./builtin/showImage");
-    const showText = require("./builtin/showText");
-    const delay = require("./builtin/delay");
-    const randomEffect = require("./builtin/randomEffect");
-    const effectGroup = require("./builtin/effectGroup");
-    const currency = require("./builtin/currency");
-    const randomRedditImage = require("./builtin/randomRedditImage");
-    const customVariable = require("./builtin/customVariable");
-    const controlMouse = require('./builtin/controlMouse');
-    const runCommand = require('./builtin/run-command');
-    const customScript = require('./builtin/customScript');
-    const clearEffects = require('./builtin/clearEffects');
-    const sequentialEffect = require("./builtin/sequentialEffect");
-    const updateRole = require("./builtin/update-role");
-    const updateVipRole = require("./builtin/update-vip-role");
-    const conditionalEffects = require("./builtin/conditional-effects/conditional-effects");
-    const loopEffects = require("./builtin/loopEffects");
-    const textToSpeech = require("./builtin/text-to-speech");
-    const deleteChatMessage = require("./builtin/delete-chat-message");
-    const stopEffectExecution = require("./builtin/stop-effect-execution");
-    const adBreak = require("./builtin/ad-break");
-    const runProgram = require("./builtin/run-program");
-    const activeUserLists = require("./builtin/activeUserLists");
-    const streamTitle = require("./builtin/stream-title");
-    const streamGame = require("./builtin/stream-game");
-    const moderatorBan = require("./builtin/moderatorBan");
-    const moderatorPurge = require("./builtin/moderatorPurge");
-    const moderatorTimeout = require("./builtin/moderatorTimeout");
-    const moderatorMod = require("./builtin/moderatorMod");
-    const clearChat = require("./builtin/clearChat");
-    const updateCounter = require("./builtin/update-counter");
-    const toggleCommand = require("./builtin/toggle-command");
-    const toggleEvent = require("./builtin/toggle-event");
-    const toggleEventSet = require("./builtin/toggle-event-set");
-    const toggleTimer = require("./builtin/toggle-timer");
-    const cooldownCommand = require("./builtin/cooldown-command");
-    const setUserMetadata = require("./builtin/set-user-metadata");
-    const shoutout = require("./builtin/shoutout");
-    const markAllActivityAcknowledged = require("./builtin/mark-all-activity-acknowledged");
-
-    // register them
-    effectManager.registerEffect(playSoundEffect); // No migration needed.
-    effectManager.registerEffect(chatEffect); // Converted for Twitch.
-    effectManager.registerEffect(chatFeedAlert);
-    effectManager.registerEffect(api); // No migration needed.
-    effectManager.registerEffect(celebration); // No migration needed.
-    effectManager.registerEffect(clips); // Converted for Twitch, needs Discord posts fixed.
-    effectManager.registerEffect(dice); // No migration needed.
-    effectManager.registerEffect(fileWriter); // No migration needed.
-    effectManager.registerEffect(html); // No migration needed.
-    effectManager.registerEffect(playVideo); // No migration needed.
-    effectManager.registerEffect(showEvents);
-    effectManager.registerEffect(controlEmulation); // No migration needed.
-    effectManager.registerEffect(showImage); // No migration needed.
-    effectManager.registerEffect(showText); // No migration needed.
-    effectManager.registerEffect(delay); // No migration needed.
-    effectManager.registerEffect(randomEffect); // No migration needed.
-    effectManager.registerEffect(effectGroup); // No migration needed.
-    effectManager.registerEffect(currency); // No migration needed.
-    effectManager.registerEffect(randomRedditImage); // No migration needed.
-    effectManager.registerEffect(customVariable); // No migration needed.
-    effectManager.registerEffect(controlMouse); // No migration needed.
-    effectManager.registerEffect(runCommand); // No migration needed.
-    effectManager.registerEffect(customScript); // No migration needed.
-    effectManager.registerEffect(clearEffects); // No migration needed.
-    effectManager.registerEffect(sequentialEffect); // No migration needed.
-    effectManager.registerEffect(updateRole); // No migration needed.
-    effectManager.registerEffect(updateVipRole);
-    effectManager.registerEffect(conditionalEffects); // No migration needed.
-    effectManager.registerEffect(loopEffects); // No migration needed.
-    effectManager.registerEffect(textToSpeech); // No migration needed.
-    effectManager.registerEffect(deleteChatMessage); // Converted for Twitch
-    effectManager.registerEffect(stopEffectExecution); // No migration needed.
-    effectManager.registerEffect(adBreak); // Converted for Twitch
-    effectManager.registerEffect(runProgram); // No migration needed.
-    effectManager.registerEffect(activeUserLists); // No migration needed.
-    effectManager.registerEffect(streamTitle);
-    effectManager.registerEffect(streamGame);
-    effectManager.registerEffect(moderatorBan);
-    effectManager.registerEffect(moderatorPurge);
-    effectManager.registerEffect(moderatorTimeout);
-    effectManager.registerEffect(moderatorMod);
-    effectManager.registerEffect(clearChat);
-    effectManager.registerEffect(updateCounter);
-    effectManager.registerEffect(toggleCommand);
-    effectManager.registerEffect(toggleEventSet);
-    effectManager.registerEffect(toggleEvent);
-    effectManager.registerEffect(toggleTimer);
-    effectManager.registerEffect(cooldownCommand);
-    effectManager.registerEffect(setUserMetadata);
-    effectManager.registerEffect(shoutout);
-    effectManager.registerEffect(markAllActivityAcknowledged);
+    effectManager.registerEffect(require("./builtin/active-user-lists")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/ad-break")); // Converted for Twitch
+    effectManager.registerEffect(require("./builtin/api")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/celebration")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/chat-feed-alert"));
+    effectManager.registerEffect(require("./builtin/chat")); // Converted for Twitch.
+    effectManager.registerEffect(require("./builtin/clear-chat"));
+    effectManager.registerEffect(require('./builtin/clear-effects')); // No migration needed.
+    effectManager.registerEffect(require('./builtin/clips')); // Converted for Twitch, needs Discord posts fixed.
+    effectManager.registerEffect(require("./builtin/conditional-effects/conditional-effects")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/control-emulation")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/cooldown-command"));
+    effectManager.registerEffect(require("./builtin/currency")); // No migration needed.
+    effectManager.registerEffect(require('./builtin/custom-script')); // No migration needed.
+    effectManager.registerEffect(require("./builtin/custom-variable")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/delay")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/delete-chat-message")); // Converted for Twitch
+    effectManager.registerEffect(require("./builtin/dice")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/effect-group")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/file-writer")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/html")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/loop-effects")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/mark-all-activity-acknowledged"));
+    effectManager.registerEffect(require("./builtin/moderator-ban"));
+    effectManager.registerEffect(require("./builtin/moderator-mod"));
+    effectManager.registerEffect(require("./builtin/moderator-purge"));
+    effectManager.registerEffect(require("./builtin/moderator-timeout"));
+    effectManager.registerEffect(require("./builtin/play-sound")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/play-video")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/random-effect")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/random-reddit-image")); // No migration needed.
+    effectManager.registerEffect(require('./builtin/run-command')); // No migration needed.
+    effectManager.registerEffect(require("./builtin/run-program")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/sequential-effect")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/set-user-metadata"));
+    effectManager.registerEffect(require("./builtin/shoutout"));
+    effectManager.registerEffect(require("./builtin/show-events"));
+    effectManager.registerEffect(require("./builtin/show-image")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/show-text")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/stop-effect-execution")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/stream-game"));
+    effectManager.registerEffect(require("./builtin/stream-title"));
+    effectManager.registerEffect(require("./builtin/text-to-speech")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/toggle-command"));
+    effectManager.registerEffect(require("./builtin/toggle-connection"));
+    effectManager.registerEffect(require("./builtin/toggle-event-set"));
+    effectManager.registerEffect(require("./builtin/toggle-event"));
+    effectManager.registerEffect(require("./builtin/toggle-timer"));
     effectManager.registerEffect(require("./builtin/update-channel-reward"));
-
-    // Deactivated for Twitch.
-    const changeScene = require('./builtin/changeScene');
-    const streamAudience = require("./builtin/stream-audience");
-    const updateControl = require("./builtin/updateControl");
-    const channelProgression = require("./builtin/channelProgression");
-    const cooldowns = require("./builtin/cooldowns");
-    const toggleConnection = require("./builtin/toggleConnection");
-    effectManager.registerEffect(changeScene);
-    effectManager.registerEffect(streamAudience);
-    effectManager.registerEffect(updateControl);
-    effectManager.registerEffect(channelProgression);
-    effectManager.registerEffect(cooldowns);
-    effectManager.registerEffect(toggleConnection);
+    effectManager.registerEffect(require("./builtin/update-counter"));
+    effectManager.registerEffect(require("./builtin/update-role")); // No migration needed.
+    effectManager.registerEffect(require("./builtin/update-vip-role"));
 };
