+++ backend/effects/builtin/clear-chat.js
