+++ backend/effects/builtin/loop-effects.js
