+++ backend/effects/builtin/moderator-timeout.js
