+++ backend/effects/builtin/show-image.js
