--- backend/effects/builtin/stream-audience.js
@@ -1,71 +1,0 @@
-"use strict";
-
-const { ControlKind, InputEvent } = require('../../interactive/constants/MixplayConstants');
-const effectModels = require("../models/effectModels");
-const { EffectTrigger } = effectModels;
-
-const { EffectCategory } = require('../../../shared/effect-constants');
-
-const streamAudience = {
-    /**
-   * The definition of the Effect
-   */
-    definition: {
-        id: "firebot:streamaudience",
-        name: "Set Stream Audience (Deprecated)",
-        description: "Set the audience rating for the stream.",
-        icon: "fad fa-users",
-        hidden: true,
-        categories: [EffectCategory.COMMON, EffectCategory.MODERATION],
-        dependencies: [],
-        triggers: effectModels.buildEffectTriggersObject(
-            [ControlKind.BUTTON, ControlKind.TEXTBOX],
-            [InputEvent.MOUSEDOWN, InputEvent.KEYDOWN, InputEvent.SUBMIT],
-            EffectTrigger.ALL
-        )
-    },
-    /**
-   * Global settings that will be available in the Settings tab
-   */
-    globalSettings: {},
-    /**
-   * The HTML template for the Options view (ie options when effect is added to something such as a button.
-   * You can alternatively supply a url to a html file via optionTemplateUrl
-   */
-    optionsTemplate: `
-    <eos-container header="Invalid Effect">
-        <div class="effect-info alert alert-info" style="margin-bottom:0;">
-            This effect was built for Mixer and no longer works.
-        </div>
-    </eos-container>
-    `,
-    /**
-   * The controller for the front end Options
-   * Port over from effectHelperService.js
-   */
-    optionsController: () => {},
-    /**
-   * When the effect is triggered by something
-   * Used to validate fields in the option template.
-   */
-    optionsValidator: effect => {
-        let errors = [];
-        return errors;
-    },
-    /**
-   * When the effect is triggered do something
-   */
-    onTriggerEvent: async () => true,
-    /**
-   * Code to run in the overlay
-   */
-    overlayExtension: {
-        dependencies: {
-            css: [],
-            js: []
-        },
-        event: {}
-    }
-};
-
-module.exports = streamAudience;
