--- backend/effects/builtin/updateControl.js
@@ -1,42 +1,0 @@
-"use strict";
-
-const { ControlKind, InputEvent } = require('../../interactive/constants/MixplayConstants');
-const effectModels = require("../models/effectModels");
-const { EffectDependency, EffectTrigger } = effectModels;
-
-const { EffectCategory } = require('../../../shared/effect-constants');
-
-const model = {
-    definition: {
-        id: "firebot:updatecontrol",
-        name: "Update Control (Deprecated)",
-        description: "Change various properties of a MixPlay Control.",
-        icon: "fad fa-bullseye-pointer",
-        hidden: true,
-        categories: [EffectCategory.ADVANCED],
-        dependencies: [EffectDependency.INTERACTIVE],
-        triggers: effectModels.buildEffectTriggersObject(
-            [ControlKind.BUTTON, ControlKind.TEXTBOX],
-            [InputEvent.MOUSEDOWN, InputEvent.KEYDOWN, InputEvent.SUBMIT],
-            EffectTrigger.ALL
-        )
-    },
-    optionsTemplate: `
-    <eos-container header="Invalid Effect">
-        <div class="effect-info alert alert-info" style="margin-bottom:0;">
-            This effect was built for Mixer and no longer works.
-        </div>
-    </eos-container>
-    `,
-    optionsController: () => {},
-    optionsValidator: () => {
-        let errors = [];
-        return errors;
-    },
-    /**
-     * When the effect is triggered by something
-     */
-    onTriggerEvent: () => true
-};
-
-module.exports = model;
